requirements.txt
django-mailer
django==1.3
fandjango
django-sentry
python-memcached
